1) close streams properly 2) replace string concatination with StringBuilder 3) make methods synchronized
